% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennr.R
\name{intersect.sets}
\alias{intersect.sets}
\title{Intersect a name list of character vectors for venn.js}
\usage{
intersect.sets(sets, labels = TRUE, iter = 1, json = list())
}
\description{
Intersect a name list of character vectors for venn.js
}
\keyword{internal}
