% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennr.R
\name{list.setdiff}
\alias{list.setdiff}
\title{Find set difference between one and multiple lists}
\usage{
list.setdiff(x, y)
}
\description{
Find set difference between one and multiple lists
}
\keyword{internal}
