% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennr.R
\name{vennr}
\alias{vennr}
\title{Area proportional venn and euler diagrams through venn.js}
\usage{
vennr(sets, raw = FALSE, labels = TRUE, font.size = "14px",
  font.family = "Helvetica,Arial,sans-serif", width = NULL,
  height = NULL)
}
\arguments{
\item{sets}{A named list of character vectors or raw input to venn.js}

\item{raw}{Use true if sets is raw input to venn.js}

\item{labels}{Use true to label overlap regions}

\item{font.size}{Font size for labels}

\item{font.family}{Font family for labels}

\item{width}{Width of plot (e.g "800px")}

\item{height}{Height of plot (e.g "600px")}
}
\description{
Area proportional venn and euler diagrams through venn.js
}
